ServiceSharingApp/requirements 2.txt

Django==3.2.8
pytz==2021.3


django-autoslug==1.9.8
django-ckeditor==6.1.0
django-crispy-forms==1.11.2
django-environ==0.4.5
django-js-asset==1.2.2
django-storages==1.11.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1

psycopg2-binary==2.8.6
Pillow==8.4.0

PyYAML==5.4.1
requests==2.25.1

psycopg2-binary==2.9.2
